\name{Observer}
\alias{Observer}
\docType{data}
\title{Observer codes for Gray Whale Southbound Migration Survey Data from 1967-2006}
\description{
Numeric observer codes and initials with names of observers.}
\usage{data(Observer)}
\format{
  A data frame with 67 records on the following 3 variables.
  \describe{
    \item{\code{ID}}{Sequential numeric key}
    \item{\code{Observer}}{Numeric code for observer used in earlier surveys}
    \item{\code{Initials}}{Initials of observers used in more recent surveys; only 2 observers have both a code and initials}
    \item{\code{Name}}{Full name of the observer}
    \item{\code{Sex}}{M/F sex of the observer}
  }
}
\keyword{datasets}
