\name{PodsizeCalibrationTable}
\alias{PodsizeCalibrationTable}
\docType{data}
\title{Pod Size Calibration Table constructed from Pod Size Calibration Data to Correct Pod Size Estimation Error}
\description{
Calibration data for pod size estimation error organized into a table with a record for each observed pod and the frequency distribution of the 
estimates by the observers. }
\usage{data(PodsizeCalibrationTable)}
\format{
  A data frame with 196 records on the following 22 variables.
  \describe{
    \item{\code{key}}{Key field constructed of Type_Year_ID from \code{\link{PodsizeCalibrationData}}}
    \item{\code{True}}{True size of pod}
    \item{\code{1}}{Number of observers that estimated the pod to be size 1}
    \item{\code{2}}{Number of observers that estimated the pod to be size 2}
    \item{\code{3}}{Number of observers that estimated the pod to be size 3}
    \item{\code{4}}{Number of observers that estimated the pod to be size 4}
    \item{\code{5}}{Number of observers that estimated the pod to be size 5}
    \item{\code{6}}{Number of observers that estimated the pod to be size 6}
    \item{\code{7}}{Number of observers that estimated the pod to be size 7}
    \item{\code{8}}{Number of observers that estimated the pod to be size 8}
    \item{\code{9}}{Number of observers that estimated the pod to be size 9}
    \item{\code{10}}{Number of observers that estimated the pod to be size 10}
    \item{\code{11}}{Number of observers that estimated the pod to be size 11}
    \item{\code{12}}{Number of observers that estimated the pod to be size 12}
    \item{\code{13}}{Number of observers that estimated the pod to be size 13}
    \item{\code{14}}{Number of observers that estimated the pod to be size 14}
    \item{\code{15}}{Number of observers that estimated the pod to be size 15}
    \item{\code{16}}{Number of observers that estimated the pod to be size 16}
    \item{\code{17}}{Number of observers that estimated the pod to be size 17}
    \item{\code{18}}{Number of observers that estimated the pod to be size 18}
    \item{\code{19}}{Number of observers that estimated the pod to be size 19}
    \item{\code{20}}{Number of observers that estimated the pod to be size 20}
    }
}
\keyword{datasets}
